organize test infrastructure
